reorganize bootstrap package, fix if statements in my-matches
