+++ src/main/java/com/gruzini/tennistico/bootstrap/CourtInitializer.java
@@ -6,11 +6,11 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Profile;
 import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
-
+import java.util.List;
 
 @Component
 @Profile("dev")
-@Order(3)
+@Order(2)
 public class CourtInitializer implements CommandLineRunner {
 
     private final CourtRepository courtRepository;
@@ -22,6 +22,15 @@ public class CourtInitializer implements CommandLineRunner {
     @Override
     public void run(String... args) throws Exception {
         final Court court1 = Court.builder()
+                .name("Krzysztof")
+                .city("Madrid")
+                .country("Spain")
+                .state("Castilla")
+                .street("Calle")
+                .phoneNumber("+00 700 777 777")
+                .zipCode("00-000")
+                .build();
+        final Court court2 = Court.builder()
                 .phoneNumber("+33 555 222 111")
                 .zipCode("55-555")
                 .street("Key Biscayne")
@@ -30,7 +39,7 @@ public class CourtInitializer implements CommandLineRunner {
                 .city("Miami")
                 .name("Crandon Park Tennis Center")
                 .build();
-        final Court court2 = Court.builder()
+        final Court court3 = Court.builder()
                 .phoneNumber("+351 54 781 781")
                 .zipCode("31-300")
                 .street("Rua do Benformoso")
@@ -39,7 +48,7 @@ public class CourtInitializer implements CommandLineRunner {
                 .city("Lisbon")
                 .name("Sport Clube Intendente")
                 .build();
-        final Court court3 = Court.builder()
+        final Court court4 = Court.builder()
                 .phoneNumber("+00 500 500")
                 .zipCode("30-30")
                 .street("Long")
@@ -48,8 +57,6 @@ public class CourtInitializer implements CommandLineRunner {
                 .city("Miami")
                 .name("Sunny Court Center")
                 .build();
-        courtRepository.save(court1);
-        courtRepository.save(court2);
-        courtRepository.save(court3);
+        courtRepository.saveAll(List.of(court1, court2, court3, court4));
     }
 }
