+++ src/main/resources/templates/my-matches.html
@@ -31,7 +31,6 @@
                                  Below you can see all your upcoming matches.
                               </div>
                               <table class="table table-hover">
-<!--                                 TODO: make modal only for player that isn't you -->
                                  <thead>
                                  <tr>
                                     <th class="gruzini-text-card-content">ID</th>
@@ -44,37 +43,35 @@
                                  </tr>
                                  </thead>
                                  <tbody>
-
                                     <tr th:each="match, iter : ${myMatches}">
                                        <td th:text="${iter.index+1}"></td>
                                        <td>
                                           <span th:if="${match.involvedPlayerHost}">
-                                             <span class="player-modal"
-                                                   th:attr="data-name=${match.host.getFullName()}, data-points=${match.host.rankingPoints},
-                                                data-age=${match.host.age}, data-gender=${match.host.gender},
-                                                data-photo=${match.host.photoUrl}"
-                                                   onclick="showModal(this.getAttribute('data-name'), this.getAttribute('data-points'),
-                                                           this.getAttribute('data-age'), this.getAttribute('data-gender'),
-                                                           this.getAttribute('data-photo'))"
-                                                   th:text="${match.host.getFullName()}">
-                                             </span>
-                                          </span>
-                                          <span th:unless="${match.involvedPlayerHost}">
                                              <span th:text="${match.host.getFullName()}"></span>
-                                          </span>
-                                          <span> vs. </span>
-                                          <span th:if="${match.involvedPlayerHost and match.guest != null}">
-                                             <span th:text="${match.guest.getFullName()}"></span>
-                                          </span>
-                                          <span th:unless="${match.involvedPlayerHost}">
-                                             <span class="player-modal"
+                                             <span> vs. </span>
+                                             <span th:if="${match.guest != null}">
+                                                <span class="player-modal"
                                                    th:attr="data-name=${match.guest.getFullName()}, data-points=${match.guest.rankingPoints},
-                                                data-age=${match.guest.age}, data-gender=${match.guest.gender},
-                                                data-photo=${match.guest.photoUrl}"
+                                                            data-age=${match.guest.age}, data-gender=${match.guest.gender},
+                                                            data-photo=${match.guest.photoUrl}"
                                                    onclick="showModal(this.getAttribute('data-name'), this.getAttribute('data-points'),
                                                            this.getAttribute('data-age'), this.getAttribute('data-gender'),
                                                            this.getAttribute('data-photo'))"
                                                    th:text="${match.guest.getFullName()}">
+                                                </span>
+                                             </span>
+                                          </span>
+                                          <span th:unless="${match.involvedPlayerHost}">
+                                             <span th:text="${match.guest.getFullName()}"></span>
+                                             <span> vs. </span>
+                                             <span class="player-modal"
+                                                   th:attr="data-name=${match.host.getFullName()}, data-points=${match.host.rankingPoints},
+                                                            data-age=${match.host.age}, data-gender=${match.host.gender},
+                                                            data-photo=${match.host.photoUrl}"
+                                                   onclick="showModal(this.getAttribute('data-name'), this.getAttribute('data-points'),
+                                                           this.getAttribute('data-age'), this.getAttribute('data-gender'),
+                                                           this.getAttribute('data-photo'))"
+                                                   th:text="${match.host.getFullName()}">
                                              </span>
                                           </span>
                                        </td>
@@ -90,12 +87,11 @@
                                           <span class="gruzini-today-sign"
                                                 th:if="${match.getStart().toLocalDate().isEqual(#temporals.createToday())}">
                                                         <span th:text="' - '+'today!'"></span>
-                                                    </span>
+                                          </span>
                                           <span th:if="${match.getStart().toLocalDate().isEqual(#temporals.createToday().plusDays(1))}"
                                                 class="gruzini-tomorrow-sign">
                                                         <span th:text="' - '+'tomorrow!'"></span>
-                                                    </span>
-
+                                          </span>
                                        </td>
                                        <td th:text="${#temporals.format(match.getStart(),'HH:mm')}"></td>
                                        <td th:text="${#temporals.format(match.getEnd(),'HH:mm')}"
@@ -113,15 +109,11 @@
                                           </a>
                                        </td>
                                     </tr>
-
                                  </tbody>
                               </table>
                            </div>
-
-
                         </div>
                         <div class="card-footer d-flex justify-content-center">
-
                         </div>
                      </div>
                   </div>
